+++ gui/elements/qt/__init__.py
@@ -4,10 +4,8 @@ wrap.
 """
 from PyQt5.QtWidgets import QApplication as QtApplication
 
-from ._layerPanel import QtLayerPanel
-from ._controlBars import QtControlBars
-from ._imageLayer import QtImageLayer
-from ._markersLayer import QtMarkersLayer
+from ._layer_list import QtLayerPanel
+from ._control_bars import QtControlBars
 from ._dimensions import QtDimensions
 from ._viewer import QtViewer
